fix: /v1/process-instance/search not tenant-aware with filter queries
The OpensearchSearchableDao initializes a search request with a tenant-aware match-all query. However, the `buildFiltering` -- and buildSorting and buildPaging -- give consumers unrestricted access to the SearchRequest. Specifically in `buildFiltering` it meant that derived classes defaulted to setting the `SearchResult.Builder.query()` property without ensuring that the query used a tenant-aware query check.

`buildFiltering` has been refactored so that it no longer allows consumers to set SearchRequest.Builder properties directly. Instead, they return a list of query terms which the OpensearchSearchableDao chains together with a tenant-aware query statement.

Ultimately a consumer can still overwrite the query property before a search by overriding the search method, for example. Further refactorings may be considered to try and tighten/restrict how a consumer can access the SearchRequest.