feat: when a job cannot be sent to the client it should be logged at INFO
When a job cannot be sent to the client, it is marked as failed and
reactivated in the broker.
This means that the job will have one less retry available. If it's the
last retry, this will result in an incident.
If the gateway log level is not set to DEBUG, then it's almost
impossible to understand for the operator what was the cause of the
incident.

This situation arised in a support case, where an incident was created
and the customer was able to reproduce the issue in a test environment
with DEBUG logs enabled.

SUPPORT-30088