test: unflake job based adhoc subprocess test
This test was flaky, because it could happen that the terminate element
command for the ad-hoc sub-process instance would be processed
completely before the activate instruction was processed. In that case
the rejection type was different: NOT_FOUND instead of INVALID_STATE.

By writing the activate instruction command directly after the
TERMINATE_ELEMENT command and using the engine with maxCommandsInBatch
set to 1, we can guarantee that the activation instruction command is
executed before the follow-up commands of the TERMINATE_ELEMENT command.

This means that the ad-hoc sub-process instance is TERMINATING rather
TERMINATED.